+++ crates/fuel-core/src/service.rs
@@ -1,14 +1,26 @@
+use self::adapters::BlockImporterAdapter;
 use crate::{
     combined_database::CombinedDatabase,
     database::Database,
-    service::adapters::P2PAdapter,
+    service::{
+        adapters::{
+            P2PAdapter,
+            PoAAdapter,
+        },
+        genesis::execute_genesis_block,
+    },
 };
+use fuel_core_poa::ports::BlockImporter;
 use fuel_core_services::{
     RunnableService,
     RunnableTask,
     ServiceRunner,
     State,
     StateWatcher,
+};
+use fuel_core_storage::{
+    transactional::AtomicView,
+    IsNotFound,
 };
 use std::net::SocketAddr;
 use tracing::warn;
@@ -20,10 +32,6 @@ pub use config::{
     VMConfig,
 };
 pub use fuel_core_services::Service as ServiceTrait;
-
-use crate::service::adapters::PoAAdapter;
-
-use self::adapters::BlockImporterAdapter;
 
 pub mod adapters;
 pub mod config;
@@ -216,7 +224,6 @@ impl Task {
             .unwrap_or_default();
         let da_block_height = 0u64.into();
         database.init(&block_height, &da_block_height)?;
-        genesis::maybe_initialize_state(&config, database.on_chain())?;
 
         // initialize sub services
         tracing::info!("Initializing sub services");
@@ -246,6 +253,16 @@ impl RunnableService for Task {
         _: &StateWatcher,
         _: Self::TaskParams,
     ) -> anyhow::Result<Self::Task> {
+        let view = self.shared.database.latest_view();
+        // check if chain is initialized
+        if let Err(err) = view.get_genesis() {
+            if err.is_not_found() {
+                let result = execute_genesis_block(&self.shared.config, &view)?;
+
+                self.shared.block_importer.commit_result(result).await?;
+            }
+        }
+
         for service in &self.services {
             service.start_and_await().await?;
         }